+++ hyperkitty/tests/lib/test_posting.py
@@ -22,6 +22,7 @@
 
 import uuid
 from smtplib import SMTPDataError
+from urllib.error import HTTPError
 
 from django.contrib.auth.models import User
 from django.core import mail
@@ -162,3 +163,14 @@ class PostingTestCase(TestCase):
         ]
         addr = posting.get_sender(self.request, self.mlist)
         self.assertTrue(isinstance(addr, str))
+
+    def test_posting_when_address_banned(self):
+        # Test that we handle exception when the user is banned.
+        with patch("hyperkitty.lib.posting.mailman.subscribe") as sub_fn:
+            sub_fn.side_effect = HTTPError(
+                400, 'Address is banned.', None, None, None)
+            with self.assertRaises(posting.PostingFailed) as cm:
+                posting.post_to_list(
+                    self.request, self.mlist, "Dummy subject", "body")
+                self.assertEqual(str(cm.exception),
+                                 'Failed to post message: Address is banned')